Added Utility for access control logic

+++ src/main/java/com/example/clubsapi/controller/ClubController.java
@@ -7,7 +7,9 @@ import com.example.clubsapi.entity.UserEntity;
 import com.example.clubsapi.exception.ResousrceNotFoundException;
 import com.example.clubsapi.repository.UserRepository;
 import com.example.clubsapi.services.impl.ClubsServiceImpl;
+import com.example.clubsapi.utiltiy.AccessControl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -25,11 +27,13 @@ public class ClubController {
 
     private ClubsServiceImpl clubsService;
     private UserRepository userRepository;
+    private AccessControl accessControl;
 
     @Autowired
-    public ClubController(ClubsServiceImpl clubsService, UserRepository userRepository) {
+    public ClubController(ClubsServiceImpl clubsService, UserRepository userRepository, AccessControl accessControl) {
         this.clubsService = clubsService;
         this.userRepository = userRepository;
+        this.accessControl = accessControl;
     }
 
     @PostMapping("/new")
@@ -43,7 +47,7 @@ public class ClubController {
 
     }
 
-    @PreAuthorize("hasRole('MODERATOR') or hasRole('USER')")
+
     @GetMapping("/allclubs")
     public ResponseEntity<?> getallClubs() {
 
@@ -55,20 +59,24 @@ public class ClubController {
     @PreAuthorize("hasRole('MODERATOR')")
     @DeleteMapping("/delete/{clubId}")
     public ResponseEntity<?> deleteClub(@PathVariable("clubId") int clubId) {
-        clubsService.deleteclub(clubId);
-        return ResponseEntity.ok().build();
 
-
+        if(accessControl.deleteClubUtility(clubId)){
+            clubsService.deleteclub(clubId);
+            return ResponseEntity.status(HttpStatus.OK).body("Club Deleted");
+        }
+        else{
+            Map<String,String> error = new HashMap<>();
+            error.put("error code ", "403");
+            error.put("Message","Your are forbidden to access the records");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(error);
+        }
     }
 
     @PreAuthorize("hasRole('MODERATOR')")
     @PatchMapping("/update/{clubId}")
     public ResponseEntity<?> updateClub(@RequestBody ClubDto clubDto, @PathVariable("clubId") int clubId) {
-        String userContext= (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        UserEntity user =userRepository.findByUsername(userContext).
-                orElseThrow(()->new ResousrceNotFoundException("User not found"));
-        if(user.getClubs().stream().toList().get(0).getId() == clubId &&
-                user.getClubs().stream().toList().get(0).getTitle().equals( clubDto.getClubName()) ){
+
+        if( accessControl.clubUpdateUtility(clubId,clubDto)){
             ClubDto cLubDto = clubsService.updateClub(clubDto, clubId);
             return ResponseEntity.ok(cLubDto);
         }
@@ -76,7 +84,7 @@ public class ClubController {
             Map<String,String> error = new HashMap<>();
             error.put("error code ", "403");
             error.put("Message","Your are forbidden to access the records");
-            return ResponseEntity.badRequest().body(error);
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(error);
         }
